pyproject.toml
[project]
name = "neuromod25"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "hydra-core>=1.3.2",
    "ipykernel>=6.29.5",
    "matplotlib>=3.7.5",
    "numpy>=1.24.4",
    "torch==2.0.1",
]

[tool.setuptools]
packages = ["nm"]
